migrated root ui under src/app
